[HOPS-644] change inodeID from int to long

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/io/hops/common/IDsGeneratorFactory.java
@@ -106,8 +106,8 @@ public class IDsGeneratorFactory {
         cacheDirectiveIdsThreshold));
   }
 
-  public int getUniqueINodeID(){
-    int id = (int)iDsGenerators.get(0).getUniqueID();
+  public long getUniqueINodeID(){
+    long id = iDsGenerators.get(0).getUniqueID();
     LOG.debug("Unique INode generated. id="+id);
     return id;
   }
